reverting back to  and simplify UI to have single view
