+++ mobile/Playground.playground/Contents.swift
@@ -1,10 +1,31 @@
 import SwiftUI
+import PlaygroundSupport
 
-// struct to display list of messages sent to s3
-struct Message: Identifiable {
-    var id = UUID()
-    var dateString: String
+struct DurationView: View {
+    @State var duration = 0.0
+    let timer = Timer.publish(every: 0.1, on: .main, in: .common).autoconnect()
+
+    var body: some View {
+        Text("\(String(format: "Duration: %.1f", duration))")
+            .onReceive(timer) { input in
+                self.duration = self.duration + 0.1
+            }
+    }
 }
-var messagesReceived: [Message] = []
-messagesReceived.append(Message(dateString: "2020"))
-print(messagesReceived)
+
+
+//struct ContentView: View {
+//    @State var timeRemaining = 10
+//    let timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
+//
+//    var body: some View {
+//        Text("\(timeRemaining)")
+//            .onReceive(timer) { _ in
+//                if self.timeRemaining > 0 {
+//                    self.timeRemaining -= 1
+//                }
+//            }
+//    }
+//}
+
+PlaygroundPage.current.setLiveView(DurationView())
