+++ mobile/mobile WatchKit Extension/HostingController.swift
@@ -12,6 +12,6 @@ import SwiftUI
 
 class HostingController: WKHostingController<ContentView> {
     override var body: ContentView {
-        return ContentView()
+        return ContentView(motion: MotionManager())
     }
 }
