+++ mobile/mobile.xcodeproj/xcuserdata/collinprather.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -17,12 +17,12 @@
 		<key>mobile WatchKit App.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>13</integer>
+			<integer>12</integer>
 		</dict>
 		<key>mobile.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>12</integer>
+			<integer>13</integer>
 		</dict>
 	</dict>
 </dict>
